Worked on creating family account, does save account to db but many missing fields and login does not work
Got form to actually save family to db, but most of the specific fields to FGP are not saving. Login also does not work, possbily due to change from email as login credential? Fixed all fields in form being required. Added some input validation for some of the family fields.